backend/requirements.txt
aniso8601==6.0.0
Click==7.0
flask
flask_cors
Flask-RESTful==0.3.7
flask_sqlalchemy
itsdangerous==1.1.0
Jinja2==2.10.1
python-jose
python-dotenv
MarkupSafe==1.1.1
psycopg2
psycopg2-binary
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
Werkzeug==0.15.4
unittest
